refactor: Unify crossruntime Page
