--- src/Uno.UI/UI/Xaml/Controls/Page/Page.wasm.cs
@@ -1,22 +1,0 @@
-﻿using System;
-using System.Drawing;
-using Uno.Extensions;
-using Uno.UI;
-using Uno.UI.DataBinding;
-using Microsoft.UI.Xaml.Media;
-using Microsoft.UI.Xaml.Shapes;
-
-namespace Microsoft.UI.Xaml.Controls
-{
-	public partial class Page
-	{
-		private void InitializeBorder()
-		{
-		}
-
-		private void UpdateBorder()
-		{
-			SetBorder(Thickness.Empty, null, default);
-		}
-	}
-}
